Replaced Postcode Enum with Type, 'Remove Tast.Result'

+++ LocalParks/LocalParks.Core/SportsClub.cs
@@ -6,7 +6,7 @@
         public string Name { get; set; }
         public Park Park { get; set; }
         public SportType Sport { get; set; }
-        public double MembershipFee { get; set; }
+        public decimal MembershipFee { get; set; }
         public int Members { get; set; }
         public string Website { get; set; }
         public string Email { get; set; }
